requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.14
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-filter==24.2
django-timezone-field==7.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
prompt_toolkit==3.0.47
python-crontab==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
